Run adaptive test assembly

Shadow is a test assembly function to perform adaptive test assembly based on the generalized shadow-test framework.

Shadow(
  config,
  constraints = NULL,
  true_theta = NULL,
  data = NULL,
  prior = NULL,
  prior_par = NULL,
  exclude = NULL,
  include_items_for_estimation = NULL,
  force_solver = FALSE,
  session = NULL,
  seed = NULL
)

# S4 method for config_Shadow
Shadow(
  config,
  constraints = NULL,
  true_theta = NULL,
  data = NULL,
  prior = NULL,
  prior_par = NULL,
  exclude = NULL,
  include_items_for_estimation = NULL,
  force_solver = FALSE,
  session = NULL,
  seed = NULL
)

Arguments

config

a config_Shadow object. Use createShadowTestConfig for this.

constraints

a constraints object representing test specifications. Use loadConstraints for this.

true_theta

(optional) true theta values to use in simulation. Either true_theta or data must be supplied.

data

(optional) a matrix containing item response data to use in simulation. Either true_theta or data must be supplied.

prior

(optional) density at each config@theta_grid to use as prior. Must be a length-nq vector or a nj * nq matrix. This overrides prior_dist and prior_par in the config. prior and prior_par cannot be used simultaneously.

prior_par

(optional) normal distribution parameters c(mean, sd) to use as prior. Must be a length-nq vector or a nj * nq matrix. This overrides prior_dist and prior_par in the config. prior and prior_par cannot be used simultaneously.

exclude

(optional) a list containing item names in $i and set names in $s to exclude from selection for each participant. The length of the list must be equal to the number of participants.

include_items_for_estimation

(optional) an examinee-wise list containing:

administered_item_pool items to include in theta estimation as item_pool object.

administered_item_resp item responses to include in theta estimation.

force_solver

if TRUE, do not check whether the solver is one of recommended solvers for complex problems (set-based assembly, partitioning). (default = FALSE)

session

(optional) used to communicate with Shiny app TestDesign.

seed

(optional) used to perform data generation internally.

Value

Shadow returns an output_Shadow_all object containing assembly results.

References

van der Linden, W. J., Reese, L. M. (1998). A model for optimal constrained adaptive testing. Applied Psychological Measurement, 22, 259-270.

van der Linden, W. J. (1998). Optimal assembly of psychological and educational tests. Applied Psychological Measurement, 22, 195-211.

van der Linden, W. J. (2000). Optimal assembly of tests with item sets. Applied Psychological Measurement, 24, 225-240.

van der Linden, W. J. (2005). Linear models for optimal test design. Springer Science & Business Media.

Examples

config <- createShadowTestConfig()
true_theta <- rnorm(1)
solution <- Shadow(config, constraints_science, true_theta)
#> Loading required namespace: progress
solution@output
#> [[1]]
#> Simulee index          : 1 
#>   True theta           : 1.02296 
#>   Final theta estimate : 1.145519 
#>   Final SE estimate    : 0.2404206 
#>    stage stimulus_index item_index item_resp item_ncat interim_theta interim_se
#> 1      1             NA        567         2         3      0.840890  0.7021182
#> 2      2             NA        914         2         3      1.145240  0.6476122
#> 3      3             NA        692         1         2      1.341791  0.6383215
#> 4      4             NA        688         1         2      1.434509  0.6207273
#> 5      5             NA        908         1         2      1.501117  0.6167497
#> 6      6             NA        481         0         2      1.299631  0.5676184
#> 7      7             NA        453         1         2      1.393980  0.5585402
#> 8      8             NA        514         1         2      1.434750  0.5569565
#> 9      9             NA          4         1         2      1.443480  0.5559597
#> 10    10             NA         81         1         2      1.448978  0.5554800
#> 11    11             NA        262         2         3      1.687413  0.5064102
#> 12    12             NA         29         2         3      1.842370  0.4883343
#> 13    13             NA        964         2         3      1.925859  0.4882781
#> 14    14             NA        802         1         2      1.993207  0.4916112
#> 15    15             NA        389         0         2      1.827605  0.4490158
#> 16    16             NA        210         1         2      1.869863  0.4503529
#> 17    17             NA        378         0         2      1.747112  0.4243044
#> 18    18             NA        421         0         2      1.604680  0.3990310
#> 19    19             NA          3         1         2      1.616850  0.3983229
#> 20    20             NA        875         1         2      1.631206  0.3983743
#> 21    21             NA        243         0         3      1.490227  0.3493464
#> 22    22             NA        739         0         3      1.348764  0.3183116
#> 23    23             NA        791         1         3      1.293530  0.2987887
#> 24    24             NA        305         1         3      1.346816  0.2878471
#> 25    25             NA         71         1         3      1.341741  0.2762194
#> 26    26             NA        101         0         3      1.271681  0.2648056
#> 27    27             NA        639         0         3      1.185350  0.2559096
#> 28    28             NA        780         1         3      1.213866  0.2509480
#> 29    29             NA        651         0         3      1.193111  0.2453854
#> 30    30             NA        946         1         3      1.145519  0.2404206
#>    theta_segment
#> 1              4
#> 2              5
#> 3              5
#> 4              5
#> 5              5
#> 6              6
#> 7              5
#> 8              5
#> 9              5
#> 10             5
#> 11             5
#> 12             6
#> 13             6
#> 14             6
#> 15             6
#> 16             6
#> 17             6
#> 18             6
#> 19             6
#> 20             6
#> 21             6
#> 22             5
#> 23             5
#> 24             5
#> 25             5
#> 26             5
#> 27             5
#> 28             5
#> 29             5
#> 30             5
#> 
#>