Run fixed-form test assembly

Static is a test assembly function to perform fixed-form test assembly based on the generalized shadow-test framework.

Static(config, constraints, force_solver = FALSE)

# S4 method for config_Static
Static(config, constraints, force_solver = FALSE)

Arguments

config
a config_Static object. Use createStaticTestConfig for this.

constraints
a constraints object representing test specifications. Use loadConstraints for this.

force_solver
if TRUE, do not check whether the solver is one of recommended solvers for doing set-based assembly. Has no effect on discrete assembly. (default = FALSE)

Value

Static returns a output_Static object containing the selected items.

References

van der Linden, W. J. (2005). Linear models for optimal test design. Springer Science & Business Media.

Examples

config_science <- createStaticTestConfig(
  list(
    method = "MAXINFO",
    target_location = c(-1, 1)
  )
)
solution <- Static(config_science, constraints_science)