Calculate log-likelihood

calcLogLikelihood is a function to calculate log-likelihood values.

calcLogLikelihood(object, theta, resp)

# S4 method for item_pool,numeric,numeric
calcLogLikelihood(object, theta, resp)

# S4 method for item_pool,numeric,matrix
calcLogLikelihood(object, theta, resp)

# S4 method for item_pool,matrix,numeric
calcLogLikelihood(object, theta, resp)

# S4 method for item_pool,matrix,matrix
calcLogLikelihood(object, theta, resp)

Arguments

object
an item_pool object.

theta
theta values to use.

resp
the response data to use.

Value

calcLogLikelihood returns values of log-likelihoods.

References

Rasch G (1960). Probabilistic models for some intelligence and attainment tests. Copenhagen: Danish Institute for Educational Research.

Lord FM (1952). “A theory of test scores.” Psychometric Monograph, 7.

Birnbaum A (1957). “Efficient design and use of tests of mental ability for various decision-making problems (Series Report No. 58-16. Project No. 7755-23).” Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum A (1958). “On the estimation of mental ability (Series Report No. 15. Project No. 7755-23).” Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum A (1958). “Further considerations of efficiency in tests of a mental ability (Series Report No. 17. Project No. 7755-23).” Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum A (1968). “Some latent trait models and their use in inferring an examinee's ability.” In Lord FM, Novick MR (eds.), Statistical Theories of Mental Test Scores, 397--479. Addison-Wesley, Reading, MA.

Masters GN (1982). “A Rasch model for partial credit scoring.” Psychometrika, 47(2), 149--174.

Andrich D (1978). “A rating formulation for ordered response categories.” Psychometrika, 43(4), 561--573.

Muraki E (1992). “A generalized partial credit model: Application of an EM algorithm.” Applied Psychological Measurement, 16(2), 159--176.

Samejima F (1969). “Estimation of latent ability using a response pattern of graded scores.” Psychometrika Monograph, 17.

Examples

j_pool   <- calcLogLikelihood(itempool_science, seq(-3, 3, 1), 0)