Calculate Fisher information (multiple items)

calc_info and calc_info_matrix are functions to calculate Fisher information. These functions are designed for multiple items.

calc_info(x, item_parm, ncat, model)

calc_info_matrix(x, item_parm, ncat, model)

Arguments

x
the theta value. This must be a column vector in matrix form for array_info_* functions.

item_parm
a matrix containing item parameters. Each row represents each item.

ncat
a vector containing the number of response categories of each item.

model
a vector indicating item models of each item, using

1: 1PL model

2: 2PL model

3: 3PL model

4: PC model

5: GPC model

6: GR model

Details

calc_info accepts a single theta value, and calc_info_matrix accepts multiple theta values.

Currently supports unidimensional models.

References

Rasch G (1960). Probabilistic models for some intelligence and attainment tests. Copenhagen: Danish Institute for Educational Research.

Lord FM (1952). “A theory of test scores.” Psychometric Monograph, 7.

Birnbaum A (1957). “Efficient design and use of tests of mental ability for various decision-making problems (Series Report No. 58-16. Project No. 7755-23).” Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum A (1958). “On the estimation of mental ability (Series Report No. 15. Project No. 7755-23).” Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum A (1958). “Further considerations of efficiency in tests of a mental ability (Series Report No. 17. Project No. 7755-23).” Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum A (1968). “Some latent trait models and their use in inferring an examinee's ability.” In Lord FM, Novick MR (eds.), Statistical Theories of Mental Test Scores, 397--479. Addison-Wesley, Reading, MA.

Masters GN (1982). “A Rasch model for partial credit scoring.” Psychometrika, 47(2), 149--174.

Andrich D (1978). “A rating formulation for ordered response categories.” Psychometrika, 43(4), 561--573.

Muraki E (1992). “A generalized partial credit model: Application of an EM algorithm.” Applied Psychological Measurement, 16(2), 159--176.

Samejima F (1969). “Estimation of latent ability using a response pattern of graded scores.” Psychometrika Monograph, 17.

Examples

# item parameters
item_parm <- matrix(c(
  1, NA,   NA,
  1,  2,   NA,
  1,  2, 0.25,
  0,  1,   NA,
  2,  0,    1,
  2,  0,    2),
  nrow = 6,
  byrow = TRUE
)

ncat  <- c(2, 2, 2, 3, 3, 3)
model <- c(1, 2, 3, 4, 5, 6)

# single theta example
x <- 0.5
calc_info(x, item_parm, ncat, model)
# same as
info_1pl(x, 1)
info_2pl(x, 1, 2)
info_3pl(x, 1, 2, 0.25)
info_pc(x, c(0, 1))
info_gpc(x, 2, c(0, 1))
info_gr(x, 2, c(0, 2))

# multiple thetas example
x <- matrix(seq(0.1, 0.5, 0.1)) # column vector in matrix form
calc_info_matrix(x, item_parm, ncat, model)
# same as
array_info_1pl(x, 1)
array_info_2pl(x, 1, 2)
array_info_3pl(x, 1, 2, 0.25)
array_info_pc(x, c(0, 1))
array_info_gpc(x, 2, c(0, 1))
array_info_gr(x, 2, c(0, 2))