Class 'constraints': a set of constraints

constraints is an S4 class for representing a set of constraints and its associated objects.

Details

See constraints-operators for object manipulation functions.

Slots

constraints

a data.frame containing the constraint specifications.

list_constraints

a list containing the constraint object representation of each constraint.

pool

the item_pool object associated with the constraints.

item_attrib

the item_attrib object associated with the constraints.

st_attrib

the st_attrib object associated with the constraints.

test_length

the test length specified in the constraints.

nv

the number of decision variables. Equals ni + ns.

ni

the number of items to search from.

ns

the number of stimulus to search from.

id

the item/stimulus ID string of each item/stimulus.

index,mat,dir,rhs

these represent MIP-format constraints. A single MIP-format constraint is associated with a value in index, a row in mat, a value in rhs, and a value in dir.

the i-th value of index represents which constraint specification in the constraints argument it was translated from.

the i-th row of mat represents LHS coefficients to use on decision variables in the i-th MIP-format constraint.

the i-th value of rhs represents RHS values to use in the i-th MIP-format constraint.

the i-th value of dir represents the imposed constraint between LHS and RHS.

set_based

TRUE if the constraint is set-based. FALSE otherwise.

item_order

the item attribute of each item to use in imposing an item order constraint, if any.

item_order_by

the name of the item attribute to use in imposing an item order constraint, if any.

stim_order

the stimulus attribute of each stimulus to use in imposing a stimulus order constraint, if any.

stim_order_by

the name of the stimulus attribute to use in imposing a stimulus order constraint, if any.

item_index_by_stimulus

a list containing item indices of each stimulus.

stimulus_index_by_item

the stimulus indices of each item.