(C++) Calculate expected scores

e_*() and array_e_*() are C++ functions for calculating expected scores.

e_1pl(x, b)

e_2pl(x, a, b)

e_m_2pl(x, a, d)

e_3pl(x, a, b, c)

e_m_3pl(x, a, d, c)

e_pc(x, b)

e_gpc(x, a, b)

e_m_gpc(x, a, d)

e_gr(x, a, b)

e_m_gr(x, a, d)

array_e_1pl(x, b)

array_e_2pl(x, a, b)

array_e_3pl(x, a, b, c)

array_e_pc(x, b)

array_e_gpc(x, a, b)

array_e_gr(x, a, b)

Arguments

x

the theta value. The number of columns should correspond to the number of dimensions. For array_*() functions, the number of theta values must correspond to the number of rows.

b, d

the difficulty parameter. b is used for unidimensional items, and d is used for multidimensional items.

a

the a-parameter.

c

the c-parameter.

Details

e_*() functions accept a single theta value, and array_p_*() functions accept multiple theta values.

Supports unidimensional and multidimensional models.

e_1pl(), array_e_1pl(): 1PL models

e_2pl(), array_e_2pl(): 2PL models

e_3pl(), array_e_3pl(): 3PL models

e_pc(), array_e_pc(): PC (partial credit) models

e_gpc(), array_e_gpc(): GPC (generalized partial credit) models

e_gr(), array_e_gr(): GR (graded response) models

e_m_2pl(), array_e_m_2pl(): multidimensional 2PL models

e_m_3pl(), array_e_m_3pl(): multidimensional 3PL models

e_m_gpc(), array_e_m_gpc(): multidimensional GPC models

e_m_gr(), array_e_m_gr(): multidimensional GR models

References

Rasch, G. (1960). Probabilistic models for some intelligence and attainment tests. Copenhagen: Danish Institute for Educational Research.

Lord, F. M. (1952). A theory of test scores (Psychometric Monograph No. 7). Richmond, VA: Psychometric Corporation.

Birnbaum, A. (1957). Efficient design and use of tests of mental ability for various decision-making problems (Series Report No. 58-16. Project No. 7755-23). Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum, A. (1958). On the estimation of mental ability (Series Report No. 15. Project No. 7755-23). Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum, A. (1958). Further considerations of efficiency in tests of a mental ability (Series Report No. 17. Project No. 7755-23). Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum, A. (1968). Some latent trait models and their use in inferring an examinee's ability. In Lord, F. M., Novick, M. R. (eds.), Statistical Theories of Mental Test Scores, 397-479. Reading, MA: Addison-Wesley.

Masters, G. N. (1982). A Rasch model for partial credit scoring. Psychometrika, 47(2), 149-174.

Andrich, D. (1978). A rating formulation for ordered response categories. Psychometrika, 43(4), 561-573.

Muraki, E. (1992). A generalized partial credit model: Application of an EM algorithm. Applied Psychological Measurement, 16(2), 159-176.

Samejima, F. (1969). Estimation of latent ability using a response pattern of graded scores. Psychometrika Monograph, 17.

Examples

x <- 0.5

e_1pl(x, 1)
#> [1] 0.3775407
e_2pl(x, 1, 2)
#> [1] 0.1824255
e_3pl(x, 1, 2, 0.25)
#> [1] 0.3868191
e_pc(x, c(0, 1))
#> [1] 1
e_gpc(x, 2, c(0, 1))
#> [1] 1
e_gr(x, 2, c(0, 2))
#> [1] 0.7784845

x <- matrix(seq(-3, 3, 1)) # three theta values, unidimensional

array_e_1pl(x, 1)
#>            [,1]
#> [1,] 0.01798621
#> [2,] 0.04742587
#> [3,] 0.11920292
#> [4,] 0.26894142
#> [5,] 0.50000000
#> [6,] 0.73105858
#> [7,] 0.88079708
array_e_2pl(x, 1, 2)
#>             [,1]
#> [1,] 0.006692851
#> [2,] 0.017986210
#> [3,] 0.047425873
#> [4,] 0.119202922
#> [5,] 0.268941421
#> [6,] 0.500000000
#> [7,] 0.731058579
array_e_3pl(x, 1, 2, 0.25)
#>           [,1]
#> [1,] 0.2550196
#> [2,] 0.2634897
#> [3,] 0.2855694
#> [4,] 0.3394022
#> [5,] 0.4517061
#> [6,] 0.6250000
#> [7,] 0.7982939
array_e_pc(x, c(0, 1))
#>            [,1]
#> [1,] 0.04912048
#> [2,] 0.13029916
#> [3,] 0.32973451
#> [4,] 0.73304361
#> [5,] 1.26695639
#> [6,] 1.67026549
#> [7,] 1.86970084
array_e_gpc(x, 2, c(0, 1))
#>            [,1]
#> [1,] 0.00247428
#> [2,] 0.01807457
#> [3,] 0.12330028
#> [4,] 0.59506841
#> [5,] 1.40493159
#> [6,] 1.87669972
#> [7,] 1.98192543
array_e_gr(x, 2, c(0, 2))
#>             [,1]
#> [1,] 0.002518021
#> [2,] 0.018321560
#> [3,] 0.121675545
#> [4,] 0.517986210
#> [5,] 1.000000000
#> [6,] 1.482013790
#> [7,] 1.878324455