(Internal) Estimate final theta

estimateFinalTheta is an internal function for estimating final theta.

estimateFinalTheta(
  o,
  j,
  position,
  augmented_item_pool,
  model_code,
  augmented_item_index,
  augmented_item_resp,
  include_items_for_estimation,
  item_parameter_sample,
  config,
  simulation_constants,
  bayesian_constants
)

Arguments

o

an output_Shadow object.

j

the numeric index of the examinee.

position

the item position, ranging from 1 to test length.

augmented_item_pool

the item_pool object, also including any extra items supplied to the simulation.

model_code

machine-sanitized item model codes for passing onto C++ functions.

augmented_item_index

item indices of items administered to this examinee, also including any extra items supplied to the simulation.

augmented_item_resp

responses for items administered to this examinee, also including any extra items supplied to the simulation.

include_items_for_estimation

an examinee-wise list containing:

administered_item_pool items to include in theta estimation as item_pool object.

administered_item_resp item responses to include in theta estimation.

item_parameter_sample

a list containing numerical samples of item parameters that reflect item parameter estimation uncertainty. The output of iparPosteriorSample goes here.

config

a config_Shadow object. Use createShadowTestConfig for this.

simulation_constants

a named list containing simulation constants.

bayesian_constants

a named list containing Bayesian constants.

Value

estimateFinalTheta returns an updated output_Shadow object.