(C++) Calculate Fisher information

info_*() and array_info_*() are functions for calculating Fisher information.

info_1pl(x, b)

info_2pl(x, a, b)

info_m_2pl(x, a, d)

dirinfo_m_2pl(x, a, d)

thisdirinfo_m_2pl(x, alpha_vec, a, d)

info_3pl(x, a, b, c)

info_m_3pl(x, a, d, c)

dirinfo_m_3pl(x, a, d, c)

thisdirinfo_m_3pl(x, alpha_vec, a, d, c)

info_pc(x, b)

info_gpc(x, a, b)

info_m_gpc(x, a, d)

dirinfo_m_gpc(x, a, d)

thisdirinfo_m_gpc(x, alpha_vec, a, d)

info_gr(x, a, b)

info_m_gr(x, a, d)

dirinfo_m_gr(x, a, d)

thisdirinfo_m_gr(x, alpha_vec, a, d)

array_info_1pl(x, b)

array_info_2pl(x, a, b)

array_info_m_2pl(x, a, d)

array_dirinfo_m_2pl(x, a, d)

array_thisdirinfo_m_2pl(x, alpha_vec, a, d)

array_info_3pl(x, a, b, c)

array_info_m_3pl(x, a, d, c)

array_dirinfo_m_3pl(x, a, d, c)

array_thisdirinfo_m_3pl(x, alpha_vec, a, d, c)

array_info_pc(x, b)

array_info_gpc(x, a, b)

array_info_m_gpc(x, a, d)

array_dirinfo_m_gpc(x, a, d)

array_thisdirinfo_m_gpc(x, alpha_vec, a, d)

array_info_gr(x, a, b)

array_info_m_gr(x, a, d)

array_dirinfo_m_gr(x, a, d)

array_thisdirinfo_m_gr(x, alpha_vec, a, d)

Arguments

x

the theta value. The number of columns should correspond to the number of dimensions. For array_*() functions, the number of theta values must correspond to the number of rows.

b, d

the difficulty parameter. b is used for unidimensional items, and d is used for multidimensional items.

a

the a-parameter.

alpha_vec

the alpha angle vector. Used for directional information in thisdirinfo_*() and array_thisdirinfo_*().

c

the c-parameter.

Details

info_*() functions accept a single theta value, and array_info_* functions accept multiple theta values.

Supports unidimensional and multidimensional models.

info_1pl(), array_info_1pl(): 1PL models

info_2pl(), array_info_2pl(): 2PL models

info_3pl(), array_info_3pl(): 3PL models

info_pc(), array_info_pc(): PC (partial credit) models

info_gpc(), array_info_gpc(): GPC (generalized partial credit) models

info_gr(), array_info_gr(): GR (graded response) models

info_m_2pl(), array_info_m_2pl(): multidimensional 2PL models

info_m_3pl(), array_info_m_3pl(): multidimensional 3PL models

info_m_gpc(), array_info_m_gpc(): multidimensional GPC models

info_m_gr(), array_info_m_gr(): multidimensional GR models

Directional information for a specific angle

thisdirinfo_m_2pl(), array_thisdirinfo_m_2pl(): multidimensional 2PL models

thisdirinfo_m_3pl(), array_thisdirinfo_m_3pl(): multidimensional 3PL models

thisdirinfo_m_gpc(), array_thisdirinfo_m_gpc(): multidimensional GPC models

thisdirinfo_m_gr(), array_thisdirinfo_m_gr(): multidimensional GR models

References

Rasch, G. (1960). Probabilistic models for some intelligence and attainment tests. Copenhagen: Danish Institute for Educational Research.

Lord, F. M. (1952). A theory of test scores (Psychometric Monograph No. 7). Richmond, VA: Psychometric Corporation.

Birnbaum, A. (1957). Efficient design and use of tests of mental ability for various decision-making problems (Series Report No. 58-16. Project No. 7755-23). Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum, A. (1958). On the estimation of mental ability (Series Report No. 15. Project No. 7755-23). Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum, A. (1958). Further considerations of efficiency in tests of a mental ability (Series Report No. 17. Project No. 7755-23). Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum, A. (1968). Some latent trait models and their use in inferring an examinee's ability. In Lord, F. M., Novick, M. R. (eds.), Statistical Theories of Mental Test Scores, 397-479. Reading, MA: Addison-Wesley.

Masters, G. N. (1982). A Rasch model for partial credit scoring. Psychometrika, 47(2), 149-174.

Andrich, D. (1978). A rating formulation for ordered response categories. Psychometrika, 43(4), 561-573.

Muraki, E. (1992). A generalized partial credit model: Application of an EM algorithm. Applied Psychological Measurement, 16(2), 159-176.

Samejima, F. (1969). Estimation of latent ability using a response pattern of graded scores. Psychometrika Monograph, 17.

Examples

x <- 0.5

info_1pl(x, 1)
#> [1] 0.2350037
info_2pl(x, 1, 2)
#> [1] 0.1491465
info_3pl(x, 1, 2, 0.25)
#> [1] 0.05275357
info_pc(x, c(0, 1))
#> [1] 0.5481372
info_gpc(x, 2, c(0, 1))
#> [1] 1.695532
info_gr(x, 2, c(0, 2))
#> [1] 0.8812569

x <- matrix(seq(0.1, 0.5, 0.1)) # three theta values, unidimensional

array_info_1pl(x, 1)
#>           [,1]
#> [1,] 0.2055003
#> [2,] 0.2139097
#> [3,] 0.2217129
#> [4,] 0.2287842
#> [5,] 0.2350037
array_info_2pl(x, 1, 2)
#>           [,1]
#> [1,] 0.1131803
#> [2,] 0.1217293
#> [3,] 0.1306057
#> [4,] 0.1397638
#> [5,] 0.1491465
array_info_3pl(x, 1, 2, 0.25)
#>            [,1]
#> [1,] 0.03177467
#> [2,] 0.03633839
#> [3,] 0.04135734
#> [4,] 0.04683233
#> [5,] 0.05275357
array_info_pc(x, c(0, 1))
#>           [,1]
#> [1,] 0.5208931
#> [2,] 0.5325674
#> [3,] 0.5411376
#> [4,] 0.5463752
#> [5,] 0.5481372
array_info_gpc(x, 2, c(0, 1))
#>          [,1]
#> [1,] 1.550026
#> [2,] 1.613100
#> [3,] 1.658760
#> [4,] 1.686325
#> [5,] 1.695532
array_info_gr(x, 2, c(0, 2))
#>           [,1]
#> [1,] 1.0154736
#> [2,] 0.9974812
#> [3,] 0.9662956
#> [4,] 0.9259869
#> [5,] 0.8812569