Operators for item pool objects

subsetItemPool is a function to create a subset of an item_pool object.

pool1 + pool2 combines two item_pool objects.

pool1 - pool2 excludes the items in pool2 from pool1.

pool1 == pool2 tests equality of the two item_pool objects.

Create a pool_cluster object.

pool_cluster1 == pool_cluster2 tests equality of the two pool_cluster objects.

subsetItemPool(pool, select = NULL)

combineItemPool(pool1, pool2)

# S3 method for item_pool
-(pool1, pool2)

# S3 method for item_pool
==(pool1, pool2)

# S4 method for item_pool,numeric
[(x, i, j, ..., drop = TRUE)

# S4 method for item_pool
c(x, ...)

makeItemPoolCluster(x, ..., names = NULL)

# S4 method for item_pool
makeItemPoolCluster(x, ..., names = NULL)

# S3 method for pool_cluster
==(pool_cluster1, pool_cluster2)

Arguments

pool
An item_pool object.

select, i, j
item indices to use in subsetting.

pool1, pool2
an item_pool object.

x, ...
item_pool objects.

drop
not used, exists for compatibility with Extract in the base R package.

names
(optional) names to use for item_pool.

pool_cluster1
a pool_cluster object.

pool_cluster2
a pool_cluster object.

Examples

subitempool <- subsetItemPool(itempool_science, 1:100)

itempool <- combineItemPool(itempool_science, itempool_reading)

subitempool <- subsetItemPool(itempool_science, 1:500)
itempool <- itempool_science - subitempool

itempool <- subsetItemPool(itempool_science, 1:500)
subitempool1 <- itempool_science - itempool
subitempool2 <- subsetItemPool(itempool_science, 501:1000)
subitempool1 == subitempool2  ## TRUE

cluster <- makeItemPoolCluster(itempool_science, itempool_reading)
cluster1 <- makeItemPoolCluster(itempool_science, itempool_reading)
cluster2 <- makeItemPoolCluster(cluster1@pools[[1]], cluster1@pools[[2]])
cluster1 == cluster2  ## TRUE