Class 'output_Shadow': adaptive assembly solution

output_Shadow is an S4 class to represent an adaptive assembly solution.

Slots

simulee_id

the numeric ID of the simulee.

true_theta

the true theta of the simulee, if was specified.

true_theta_segment

the segment number of the true theta.

final_theta_est

final theta estimate.

final_se_est

the standard error of final_theta_est.

administered_item_index

item IDs administered at each position.

administered_item_resp

item responses from the simulee at each position.

administered_item_ncat

the number of categories of each administered item.

administered_stimulus_index

stimulus IDs administered at each position.

shadow_test_refreshed

TRUE indicates the shadow test was refreshed for the position.

shadow_test_feasible

TRUE indicates the MIP was feasible with all constraints.

solve_time

elapsed time in running the solver at each position.

interim_theta_est

interim theta estimates at each position.

interim_se_est

the standard error of the interim estimate at each position.

theta_segment_index

segment numbers of interim theta estimates.

prior

prior distribution, if was specified.

prior_par

prior parameters, if were specified.

posterior

the posterior distribution after completing test.

posterior_sample

foo.

likelihood

the likelihood distribution after completing test.

shadow_test

the list containing the item IDs within the shadow test used in each position.

max_cat_pool

the maximum number of response categories the item pool had.