(C++) Calculate a theta estimate using EB (Empirical Bayes) method

theta_EB_single() and theta_EB() are functions to calculate a theta estimate using EB (Empirical Bayes) method.

theta_EB(
  nx,
  theta_init,
  theta_prop,
  item_parm,
  resp,
  ncat,
  model,
  prior,
  prior_parm
)

theta_EB_single(
  nx,
  theta_init,
  theta_prop,
  item_parm,
  resp,
  ncat,
  model,
  prior,
  prior_parm
)

Arguments

nx

the number of MCMC draws.

theta_init

the initial estimate to use.

theta_prop

the SD of the proposal distribution.

item_parm

a matrix containing item parameters. Each row should represent an item.

resp

a vector containing responses on each item.

ncat

a vector containing the number of response categories of each item.

model

a vector indicating item models of each item, using

1: 1PL model

2: 2PL model

3: 3PL model

4: PC model

5: GPC model

6: GR model

prior

an integer indicating the type of prior distribution, using

1: normal distribution

2: uniform distribution

prior_parm

a vector containing parameters for the prior distribution.

Details

theta_EB_single() is designed for one item, and theta_EB() is designed for multiple items.

Currently supports unidimensional models.

Examples

# item parameters
item_parm <- matrix(c(
  1, NA,   NA,
  1,  2,   NA,
  1,  2, 0.25,
  0,  1,   NA,
  2,  0,    1,
  2,  0,    2),
  nrow = 6,
  byrow = TRUE
)

ncat  <- c(2, 2, 2, 3, 3, 3)
model <- c(1, 2, 3, 4, 5, 6)
resp  <- c(0, 1, 0, 1, 0, 1)

nx <- 100
theta_init <- 0
theta_prop <- 1.0
set.seed(1)
theta_EB_single(nx, theta_init, theta_prop, item_parm[1, ], resp[1], ncat[1], model[1], 1, c(0, 1))
#>               [,1]
#>   [1,] -0.62645381
#>   [2,] -0.62645381
#>   [3,] -0.62645381
#>   [4,] -0.62645381
#>   [5,] -0.13902476
#>   [6,] -0.14479193
#>   [7,] -0.45018032
#>   [8,] -1.24918957
#>   [9,] -1.87043015
#>  [10,] -2.16964527
#>  [11,] -2.21457887
#>  [12,] -2.21457887
#>  [13,] -1.39335768
#>  [14,] -1.61762556
#>  [15,] -0.83548926
#>  [16,] -0.03129975
#>  [17,]  0.58852599
#>  [18,]  0.58852599
#>  [19,] -0.88222639
#>  [20,] -0.83550022
#>  [21,]  0.52317933
#>  [22,]  0.08986902
#>  [23,]  0.03606398
#>  [24,]  0.03606398
#>  [25,] -0.35822598
#>  [26,]  0.88007812
#>  [27,]  0.88007812
#>  [28,]  1.44082421
#>  [29,]  1.44082421
#>  [30,]  0.27425367
#>  [31,] -0.43324149
#>  [32,] -0.43324149
#>  [33,] -0.54558770
#>  [34,] -0.27945034
#>  [35,] -0.89147673
#>  [36,] -0.89147673
#>  [37,] -0.89147673
#>  [38,] -0.27323344
#>  [39,] -0.27323344
#>  [40,] -0.27323344
#>  [41,] -0.27323344
#>  [42,] -1.21388260
#>  [43,] -1.18588044
#>  [44,] -0.94361696
#>  [45,] -0.94361696
#>  [46,] -0.69520431
#>  [47,] -0.69520431
#>  [48,] -0.43786594
#>  [49,]  0.17286042
#>  [50,]  0.17286042
#>  [51,]  0.46430665
#>  [52,]  0.34655305
#>  [53,]  0.42089438
#>  [54,] -0.37619515
#>  [55,] -0.51137376
#>  [56,] -0.73088939
#>  [57,] -0.13694320
#>  [58,]  0.86004366
#>  [59,]  0.55585974
#>  [60,]  0.55585974
#>  [61,]  0.01333970
#>  [62,]  0.02171066
#>  [63,]  0.02171066
#>  [64,]  0.14142831
#>  [65,] -1.13516390
#>  [66,] -0.90614431
#>  [67,] -1.37954495
#>  [68,] -1.37954495
#>  [69,] -1.37954495
#>  [70,] -1.37954495
#>  [71,]  0.38774232
#>  [72,]  0.38774232
#>  [73,]  0.38774232
#>  [74,] -0.02477757
#>  [75,] -0.48642230
#>  [76,] -0.45894696
#>  [77,] -0.66632771
#>  [78,] -0.66632771
#>  [79,] -0.94544101
#>  [80,] -0.80738830
#>  [81,] -0.80738830
#>  [82,] -1.87608101
#>  [83,] -2.05563754
#>  [84,] -0.47554586
#>  [85,] -0.54911026
#>  [86,] -0.54911026
#>  [87,] -0.87338053
#>  [88,] -0.87338053
#>  [89,] -0.34188434
#>  [90,]  0.34748836
#>  [91,] -1.18896147
#>  [92,] -2.14585335
#>  [93,] -2.14585335
#>  [94,] -1.40457704
#>  [95,] -0.22799373
#>  [96,] -1.31449678
#>  [97,] -1.31449678
#>  [98,] -1.31449678
#>  [99,] -1.29710116
#> [100,] -1.29710116
theta_EB(nx, theta_init, theta_prop, item_parm, resp, ncat, model, 1, c(0, 1))
#>               [,1]
#>   [1,]  0.45018710
#>   [2,]  0.35366218
#>   [3,] -0.57569997
#>   [4,] -0.82788313
#>   [5,]  0.17214567
#>   [6,]  0.15961633
#>   [7,]  0.15961633
#>   [8,]  0.15961633
#>   [9,]  0.15961633
#>  [10,]  0.15961633
#>  [11,]  0.15961633
#>  [12,]  0.15961633
#>  [13,]  0.01521672
#>  [14,]  0.01521672
#>  [15,]  0.12101909
#>  [16,] -0.49393118
#>  [17,] -0.49393118
#>  [18,] -0.49393118
#>  [19,] -0.49393118
#>  [20,] -0.49393118
#>  [21,] -0.49393118
#>  [22,] -0.49393118
#>  [23,] -0.49393118
#>  [24,] -0.58438730
#>  [25,]  0.88020001
#>  [26,]  0.67885880
#>  [27,] -0.24725069
#>  [28,] -0.08546206
#>  [29,] -0.08546206
#>  [30,] -0.08546206
#>  [31,] -0.08546206
#>  [32,] -0.06782551
#>  [33,]  0.34414920
#>  [34,] -0.70156845
#>  [35,] -0.70156845
#>  [36,] -0.10772750
#>  [37,] -0.10772750
#>  [38,] -0.10772750
#>  [39,]  0.43359983
#>  [40,] -0.06777849
#>  [41,] -0.23215432
#>  [42,] -0.23215432
#>  [43,] -0.23215432
#>  [44,]  0.94000423
#>  [45,]  0.63126366
#>  [46,]  0.63126366
#>  [47,]  0.58655452
#>  [48,]  0.04479416
#>  [49,]  0.04479416
#>  [50,]  0.04479416
#>  [51,]  0.04479416
#>  [52,]  0.04479416
#>  [53,]  0.24198760
#>  [54,]  0.24198760
#>  [55,]  0.24198760
#>  [56,]  0.24198760
#>  [57,]  0.18226432
#>  [58,]  0.18226432
#>  [59,]  0.18226432
#>  [60,]  0.02108582
#>  [61,]  0.02108582
#>  [62,]  0.02108582
#>  [63,]  0.02108582
#>  [64,]  0.02108582
#>  [65,] -0.52370492
#>  [66,] -0.62520837
#>  [67,]  0.39525554
#>  [68,]  0.39525554
#>  [69,]  0.32560072
#>  [70,]  0.32560072
#>  [71,]  1.47182908
#>  [72,]  1.47182908
#>  [73,]  1.47182908
#>  [74,]  1.47182908
#>  [75,]  1.08259190
#>  [76,]  1.08259190
#>  [77,]  1.08259190
#>  [78,] -0.80743524
#>  [79,] -0.80743524
#>  [80,] -0.80743524
#>  [81,] -0.80743524
#>  [82,] -1.25082005
#>  [83,]  1.39834683
#>  [84,]  1.39834683
#>  [85,] -0.89077715
#>  [86,] -1.18968646
#>  [87,] -0.26988278
#>  [88,] -0.26988278
#>  [89,] -0.26988278
#>  [90,]  0.09957632
#>  [91,]  0.09957632
#>  [92,]  0.09957632
#>  [93,]  0.09957632
#>  [94,]  0.09957632
#>  [95,]  0.47594661
#>  [96,]  0.26206066
#>  [97,]  0.26206066
#>  [98,]  0.26206066
#>  [99,]  0.26206066
#> [100,]  0.24968789