Run adaptive test assembly

Shadow is a test assembly function for performing adaptive test assembly based on the generalized shadow-test framework.

Shadow(
  config,
  constraints = NULL,
  true_theta = NULL,
  data = NULL,
  prior = NULL,
  prior_par = NULL,
  exclude = NULL,
  include_items_for_estimation = NULL,
  force_solver = FALSE,
  session = NULL,
  seed = NULL
)

# S4 method for config_Shadow
Shadow(
  config,
  constraints = NULL,
  true_theta = NULL,
  data = NULL,
  prior = NULL,
  prior_par = NULL,
  exclude = NULL,
  include_items_for_estimation = NULL,
  force_solver = FALSE,
  session = NULL,
  seed = NULL
)

Arguments

config

a config_Shadow object. Use createShadowTestConfig for this.

constraints

a constraints object representing test specifications. Use loadConstraints for this.

true_theta

(optional) true theta values to use in simulation. Either true_theta or data must be supplied.

data

(optional) a matrix containing item response data to use in simulation. Either true_theta or data must be supplied.

prior

(optional) density at each config@theta_grid to use as prior. Must be a length-nq vector or a nj * nq matrix. This overrides prior_dist and prior_par in the config. prior and prior_par cannot be used simultaneously.

prior_par

(optional) normal distribution parameters c(mean, sd) to use as prior. Must be a length-nq vector or a nj * nq matrix. This overrides prior_dist and prior_par in the config. prior and prior_par cannot be used simultaneously.

exclude

(optional) a list containing item names in $i and set names in $s to exclude from selection for each participant. The length of the list must be equal to the number of participants.

include_items_for_estimation

(optional) an examinee-wise list containing:

administered_item_pool items to include in theta estimation as item_pool object.

administered_item_resp item responses to include in theta estimation.

force_solver

if TRUE, do not check whether the solver is one of recommended solvers for complex problems (set-based assembly, partitioning). (default = FALSE)

session

(optional) used to communicate with Shiny app TestDesign.

seed

(optional) used to perform data generation internally.

Value

Shadow returns an output_Shadow_all object containing assembly results.

References

van der Linden, W. J., Reese, L. M. (1998). A model for optimal constrained adaptive testing. Applied Psychological Measurement, 22, 259-270.

van der Linden, W. J. (1998). Optimal assembly of psychological and educational tests. Applied Psychological Measurement, 22, 195-211.

van der Linden, W. J. (2000). Optimal assembly of tests with item sets. Applied Psychological Measurement, 24, 225-240.

van der Linden, W. J. (2005). Linear models for optimal test design. Springer Science & Business Media.

Examples

config <- createShadowTestConfig()
true_theta <- rnorm(1)
solution <- Shadow(config, constraints_science, true_theta)
#> Loading required namespace: progress
solution@output
#> [[1]]
#> Simulee index          : 1 
#>   True theta           : -1.283625 
#>   Final theta estimate : -1.036614 
#>   Final SE estimate    : 0.2504388 
#>    stage stimulus_index item_index item_resp item_ncat interim_theta interim_se
#> 1      1             NA        567         0         3    -0.5924506  0.7490406
#> 2      2             NA        795         0         3    -1.3194239  0.6560083
#> 3      3             NA        847         2         3    -1.0499667  0.5307458
#> 4      4             NA        327         0         3    -1.2930247  0.4748382
#> 5      5             NA        940         2         3    -1.0504206  0.4262484
#> 6      6             NA        587         1         3    -0.9886108  0.4019669
#> 7      7             NA        900         0         3    -1.1149921  0.3872746
#> 8      8             NA        967         0         2    -1.2086200  0.3764029
#> 9      9             NA          4         1         2    -1.1667104  0.3702666
#> 10    10             NA         92         1         2    -1.0553680  0.3656374
#> 11    11             NA         56         2         3    -0.8931584  0.3352800
#> 12    12             NA        290         0         4    -0.9808338  0.3223815
#> 13    13             NA        291         2         3    -0.8499633  0.3107308
#> 14    14             NA        664         0         2    -0.8976779  0.3036105
#> 15    15             NA        428         1         2    -0.8547658  0.2972081
#> 16    16             NA        538         0         2    -0.9233031  0.2923416
#> 17    17             NA        935         0         2    -0.9766809  0.2896060
#> 18    18             NA        517         1         2    -0.9496620  0.2869237
#> 19    19             NA        421         0         2    -0.9814084  0.2850386
#> 20    20             NA          3         0         2    -1.0141284  0.2824396
#> 21    21             NA        263         0         3    -1.0890228  0.2755081
#> 22    22             NA        563         0         3    -1.1121603  0.2713369
#> 23    23             NA        865         0         3    -1.1371820  0.2684389
#> 24    24             NA        946         0         3    -1.1591751  0.2657726
#> 25    25             NA        680         1         3    -1.0972273  0.2597884
#> 26    26             NA        662         0         2    -1.1132280  0.2583655
#> 27    27             NA        115         1         2    -1.0666776  0.2551314
#> 28    28             NA        629         1         2    -1.0426177  0.2529637
#> 29    29             NA        679         1         2    -1.0196799  0.2511977
#> 30    30             NA        352         0         2    -1.0366137  0.2504388
#>    theta_segment
#> 1              4
#> 2              3
#> 3              3
#> 4              3
#> 5              3
#> 6              3
#> 7              3
#> 8              3
#> 9              3
#> 10             3
#> 11             3
#> 12             3
#> 13             3
#> 14             3
#> 15             3
#> 16             3
#> 17             3
#> 18             3
#> 19             3
#> 20             3
#> 21             3
#> 22             3
#> 23             3
#> 24             3
#> 25             3
#> 26             3
#> 27             3
#> 28             3
#> 29             3
#> 30             3
#> 
#>