Run adaptive test assembly

Shadow is a test assembly function for performing adaptive test assembly based on the generalized shadow-test framework.

Shadow(
  config,
  constraints = NULL,
  true_theta = NULL,
  data = NULL,
  prior = NULL,
  prior_par = NULL,
  exclude = NULL,
  include_items_for_estimation = NULL,
  force_solver = FALSE,
  session = NULL,
  seed = NULL
)

# S4 method for config_Shadow
Shadow(
  config,
  constraints = NULL,
  true_theta = NULL,
  data = NULL,
  prior = NULL,
  prior_par = NULL,
  exclude = NULL,
  include_items_for_estimation = NULL,
  force_solver = FALSE,
  session = NULL,
  seed = NULL
)

Arguments

config

a config_Shadow object. Use createShadowTestConfig for this.

constraints

a constraints object representing test specifications. Use loadConstraints for this.

true_theta

(optional) true theta values to use in simulation. Either true_theta or data must be supplied.

data

(optional) a matrix containing item response data to use in simulation. Either true_theta or data must be supplied.

prior

(optional) density at each config@theta_grid to use as prior. Must be a length-nq vector or a nj * nq matrix. This overrides prior_dist and prior_par in the config. prior and prior_par cannot be used simultaneously.

prior_par

(optional) normal distribution parameters c(mean, sd) to use as prior. Must be a length-nq vector or a nj * nq matrix. This overrides prior_dist and prior_par in the config. prior and prior_par cannot be used simultaneously.

exclude

(optional) a list containing item names in $i and set names in $s to exclude from selection for each participant. The length of the list must be equal to the number of participants.

include_items_for_estimation

(optional) an examinee-wise list containing:

administered_item_pool items to include in theta estimation as item_pool object.

administered_item_resp item responses to include in theta estimation.

force_solver

if TRUE, do not check whether the solver is one of recommended solvers for complex problems (set-based assembly, partitioning). (default = FALSE)

session

(optional) used to communicate with Shiny app TestDesign.

seed

(optional) used to perform data generation internally.

Value

Shadow returns an output_Shadow_all object containing assembly results.

References

van der Linden, W. J., Reese, L. M. (1998). A model for optimal constrained adaptive testing. Applied Psychological Measurement, 22, 259-270.

van der Linden, W. J. (1998). Optimal assembly of psychological and educational tests. Applied Psychological Measurement, 22, 195-211.

van der Linden, W. J. (2000). Optimal assembly of tests with item sets. Applied Psychological Measurement, 24, 225-240.

van der Linden, W. J. (2005). Linear models for optimal test design. Springer Science & Business Media.

Examples

config <- createShadowTestConfig()
true_theta <- rnorm(1)
solution <- Shadow(config, constraints_science, true_theta)
#> Loading required namespace: progress
solution@output
#> [[1]]
#> Simulee index          : 1 
#>   True theta           : -1.593115 
#>   Final theta estimate : -1.507687 
#>   Final SE estimate    : 0.2737895 
#>    stage stimulus_index item_index item_resp item_ncat interim_theta interim_se
#> 1      1             NA        567         0         3    -0.5924506  0.7490406
#> 2      2             NA        795         1         3    -0.6443670  0.5742247
#> 3      3             NA        327         0         3    -1.0285413  0.5224673
#> 4      4             NA        940         0         3    -1.3375843  0.4973221
#> 5      5             NA        847         0         3    -1.6529421  0.4631972
#> 6      6             NA         89         2         3    -1.2656084  0.3907258
#> 7      7             NA        587         0         3    -1.3561868  0.3784217
#> 8      8             NA        797         0         2    -1.5061935  0.3740112
#> 9      9             NA        238         1         2    -1.4566129  0.3611288
#> 10    10             NA          4         1         2    -1.4125294  0.3551927
#> 11    11             NA        448         0         3    -1.5413589  0.3479523
#> 12    12             NA        428         1         2    -1.4572391  0.3372593
#> 13    13             NA        615         0         2    -1.5307616  0.3318353
#> 14    14             NA        538         0         2    -1.5838347  0.3273118
#> 15    15             NA        600         0         2    -1.6452496  0.3237429
#> 16    16             NA        472         1         3    -1.6617509  0.3187009
#> 17    17             NA        420         1         2    -1.6264356  0.3108116
#> 18    18             NA        937         1         2    -1.5821548  0.3058787
#> 19    19             NA        421         0         2    -1.6067072  0.3046031
#> 20    20             NA          3         0         2    -1.6294590  0.3027910
#> 21    21             NA        263         1         3    -1.5330838  0.2890926
#> 22    22             NA        865         0         3    -1.5504691  0.2868506
#> 23    23             NA        549         0         2    -1.5836751  0.2858473
#> 24    24             NA        946         0         3    -1.5978845  0.2839542
#> 25    25             NA        184         1         2    -1.5601280  0.2810342
#> 26    26             NA        680         0         3    -1.5736362  0.2796059
#> 27    27             NA         36         1         2    -1.5252218  0.2752969
#> 28    28             NA        352         1         2    -1.4912363  0.2741466
#> 29    29             NA        194         0         2    -1.5035741  0.2738295
#> 30    30             NA        998         0         2    -1.5076871  0.2737895
#>    theta_segment
#> 1              4
#> 2              3
#> 3              3
#> 4              3
#> 5              3
#> 6              2
#> 7              3
#> 8              3
#> 9              2
#> 10             3
#> 11             3
#> 12             2
#> 13             3
#> 14             2
#> 15             2
#> 16             2
#> 17             2
#> 18             2
#> 19             2
#> 20             2
#> 21             2
#> 22             2
#> 23             2
#> 24             2
#> 25             2
#> 26             2
#> 27             2
#> 28             2
#> 29             3
#> 30             2
#> 
#>