Compute expected a posteriori estimates of theta

eap is a function for computing expected a posteriori estimates of theta.

eap(
  object,
  select = NULL,
  resp,
  theta_grid = seq(-4, 4, 0.1),
  prior = rep(1/81, 81)
)

# S4 method for item_pool
eap(
  object,
  select = NULL,
  resp,
  theta_grid = seq(-4, 4, 0.1),
  prior = rep(1/81, 81)
)

EAP(object, select = NULL, prior, reset_prior = FALSE)

# S4 method for test
EAP(object, select = NULL, prior, reset_prior = FALSE)

# S4 method for test_cluster
EAP(object, select = NULL, prior, reset_prior = FALSE)

Arguments

object

an item_pool object.

select

(optional) if item indices are supplied, only the specified items are used.

resp

item response on all (or selected) items in the object argument. Can be a vector, a matrix, or a data frame. length(resp) or ncol(resp) must be equal to the number of all (or selected) items.

theta_grid

the theta grid to use as quadrature points. (default = seq(-4, 4, .1))

prior

a prior distribution, a numeric vector for a common prior or a matrix for individualized priors. (default = rep(1 / 81, 81))

reset_prior

used for test_cluster objects. If TRUE, reset the prior distribution for each test object.

Value

eap returns a list containing estimated values.

th theta value.

se standard error.

Examples

eap(itempool_fatigue, resp = resp_fatigue_data[10, ])
#> $th
#> [1] -0.1115028
#> 
#> $se
#> [1] 0.06188415
#> 
eap(itempool_fatigue, select = 1:20, resp = resp_fatigue_data[10, 1:20])
#> $th
#> [1] -0.1583389
#> 
#> $se
#> [1] 0.1290421
#>