Class 'output_Shadow': adaptive assembly solution for one simulee

output_Shadow is an S4 class to represent the adaptive assembly solution for one simulee.

Slots

simulee_id

the numeric ID of the simulee.

true_theta

the true theta of the simulee, if was specified.

true_theta_segment

the segment number of the true theta.

final_theta_est

final theta estimate.

final_se_est

the standard error of final_theta_est.

administered_item_index

item IDs administered at each position.

administered_item_resp

item responses from the simulee at each position.

administered_item_ncat

the number of categories of each administered item.

administered_stimulus_index

stimulus IDs administered at each position.

shadow_test_refreshed

TRUE indicates the shadow test was refreshed for the position.

shadow_test_feasible

TRUE indicates the MIP was feasible with all constraints.

solve_time

elapsed time in running the solver at each position.

initial_theta_est

initial theta estimate.

interim_theta_est

interim theta estimates at each position.

interim_se_est

the standard error of the interim estimate at each position.

theta_segment_index

segment numbers of interim theta estimates.

prior

prior distribution, if was specified.

prior_par

prior parameters, if were specified.

posterior

the posterior distribution after completing test.

posterior_sample

posterior samples of interim theta before the estimation of final theta. mean(posterior_sample) == interim_theta_est[test_length] holds.

likelihood

the likelihood distribution after completing test.

shadow_test

the list containing the item IDs within the shadow test used in each position.

max_cat_pool

the maximum number of response categories the item pool had.

ni_pool

the total number of items the item pool had.

ns_pool

the total number of stimuli the item pool had.

test_length_constraints

the test length constraint used in assembly.

set_based

whether the item pool was set-based.

item_index_by_stimulus

the list of items by each stimulus the item pool had.