Extension of plot() for objects in TestDesign package

Extension of plot() for objects in TestDesign package

Plot from objects in the TestDesign package.

# S4 method for item_pool
plot(
  x,
  y,
  theta = seq(-3, 3, 0.1),
  type = "info",
  info_type = "FISHER",
  plot_sum = TRUE,
  select = NULL,
  color = "blue",
  file_pdf = NULL,
  width = 7,
  height = 6,
  mfrow = c(1, 1),
  ...
)

# S4 method for output_Static
plot(
  x,
  y,
  theta = seq(-3, 3, 0.1),
  type = NULL,
  info_type = "FISHER",
  plot_sum = TRUE,
  select = NULL,
  color = "blue",
  file_pdf = NULL,
  width = 7,
  height = 6,
  mfrow = c(2, 4),
  ...
)

# S4 method for constraints
plot(
  x,
  y,
  theta = seq(-3, 3, 0.1),
  type = "info",
  info_type = "FISHER",
  plot_sum = TRUE,
  select = NULL,
  color = "black",
  file_pdf = NULL,
  width = 7,
  height = 6,
  mfrow = c(1, 1),
  ...
)

Arguments

x
accepts the following signatures:

item_pool:

constraints:

output_Static:

output_Shadow:

y
not used, exists for compatibility with plot in the base R package.

theta
the theta grid to use in plotting. (default = seq(-3, 3, .1))

type
the type of plot. info to plot information, score to plot expected scores, shadow to plot shadow test chart, audit to plot audit trail, and exposure to plot exposure rates.

info_type
the type of information. Currently only accepts FISHER. (default = FISHER)

plot_sum
When 'object' is an item_pool object, if TRUE then draw pool-level information, and if FALSE draw item-level information for every item in the pool.

select
A vector of indices identifying the items to subset, for when 'object' is an item_pool object.

color
the color of the curve.

file_pdf
if supplied a filename, save as a PDF file.

width
the width of graphics device.

height
the height of graphics device.

mfrow
multipanel configurations in c(nrow, ncol).

...
arguments to pass onto plot.

Examples

## Plot item information of a pool
subitempool <- subsetItemPool(itempool_science, 1:8)
plot(subitempool)
plot(itempool_science, select = 1:8)

## Plot expected score of a pool
plot(subitempool, type = "score")
plot(itempool_science, type = "score", select = 1:8)

## Plot assembly results from Static()
config <- createStaticTestConfig()
solution <- Static(config, constraints_science)
plot(solution)

## Plot attainable information range from constraints
plot(constraints_science)